refactor(transform): #205 move typed-value conversion out of transformer.go to keep the 500-line cap Signed-off-by: Ruoshi Li <iceboundrock@gmail.com>
